Adding HealthCheckTag + refactoring
